+++ PCB.h
@@ -6,17 +6,19 @@
 */
 
 
-#include <unistd.h>
+
 #ifndef PCB_H
 #define PCB_H
 
+#include <unistd.h>
+#include "constants.h"
 struct ProcessControlBlock
 {
     int occupied;     
     pid_t pid;        
-    unsigned long long nanoSeconds;
+    long nanoSeconds;
     int blocked;
-    unsigned long long blockedTime;    
+    long blockedTime;    
 };
 
 class PCB
@@ -30,12 +32,12 @@ class PCB
         void PrintPCB();
 
         //  returns 0 if successful, -1 if PCB is full
-        int addProcess(pid_t pid, unsigned long long nanoSeconds);
+        int addProcess(pid_t pid, long nanoSeconds);
         //  returns 0 if successful, -1 if PCB does not contain the process
         int removeProcess(pid_t pid);
         int getNumProcesses();
 
-        void setBlocked(pid_t pid, unsigned long long blockedTime);
+        void setBlocked(pid_t pid, long blockedTime);
 
         void nextProcess();
